test: split AIGW well-known coverage
