+++ tests2/core/aigw-wellknown-translate.test.ts
@@ -1,57 +1,13 @@
 // v2-native — NOT a migrated legacy test. Listed in tests-map.json `v2Native`.
-//
-// Unit coverage for well-known-driven AIGW discovery & translation
-// (src/server/agent/aigw-manager.ts). Verifies the pure translator against a
-// captured `.well-known/opencode` fixture, the writeAigwModelsJson emission of
-// authoritative per-model api/baseUrl, the fallback option-1 OpenAI-responses
-// routing (with NO well-known probe under skip/no-external guards), and
-// default-model seeding.
+// Pure AIGW well-known translation, filtering, collision, pricing, and default-seeding coverage.
 
-import { guardProcessEnv } from "./helpers/env-guard.js";
-guardProcessEnv();
-
-import { describe, it, beforeEach, afterEach } from "vitest";
+import { describe, it } from "vitest";
 import assert from "node:assert/strict";
-import http from "node:http";
-import dnsCallback from "node:dns";
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
-import { fileURLToPath, pathToFileURL } from "node:url";
-
-const HERE = path.dirname(fileURLToPath(import.meta.url));
-const FIXTURE_PATH = path.join(HERE, "fixtures", "wellknown-opencode.json");
-
-const {
+import {
+	GATEWAY, byId, collectAigwProviderDnsHosts, createAigwGuardedLookup, loadFixture,
+	normalizeAigwPricing, normalizeWellKnownCost, seedDefaultModelsFromWellKnown,
 	translateWellKnown,
-	writeAigwModelsJson,
-	discoverAigwModels,
-	configureAigw,
-	seedDefaultModelsFromWellKnown,
-	configureAigwRuntimeFlags,
-	normalizeAigwModelString,
-	fetchWellKnownConfig,
-	normalizeAigwPricing,
-	normalizeWellKnownCost,
-	createAigwGuardedLookup,
-	collectAigwProviderDnsHosts,
-	filterValidatedProviderUrls,
-	getAigwProviderDnsGuardHosts,
-	replaceAigwProviderDnsGuardHosts,
-	removeAigw,
-	writeAigwDnsGuardExtension,
-} = await import("../../src/server/agent/aigw-manager.ts");
-const { resetAgentDirStateForTests } = await import("../../src/server/bobbit-dir.js");
-
-function loadFixture(): any {
-	return JSON.parse(fs.readFileSync(FIXTURE_PATH, "utf-8"));
-}
-
-const GATEWAY = "http://aigw-local.t3.zone";
-
-function byId(models: any[], id: string): any {
-	return models.find((m) => (m.wireId ?? m.id) === id || m.id === id);
-}
+} from "./helpers/aigw-wellknown-test-helpers.js";
 
 describe("translateWellKnown — provider → pi-ai api mapping", () => {
 	it("routes @ai-sdk/openai to openai-responses on its own /openai/v1 baseURL with bare ids", () => {
@@ -134,164 +90,6 @@ describe("translateWellKnown — filters and metadata", () => {
 		// cost without cache_* → heuristic cache ratios applied off input.
 		assert.deepEqual(luna.cost, { input: 0.5, output: 4, cacheRead: 0.05, cacheWrite: 0.625 });
 		assert.deepEqual(luna.input, ["text"]);
-	});
-});
-
-describe("writeAigwModelsJson — authoritative per-model api/baseUrl", () => {
-	let tmpAgentDir: string;
-	let prevAgentDir: string | undefined;
-
-	beforeEach(() => {
-		prevAgentDir = process.env.BOBBIT_AGENT_DIR;
-		tmpAgentDir = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-wk-models-"));
-		process.env.BOBBIT_AGENT_DIR = tmpAgentDir;
-		resetAgentDirStateForTests();
-	});
-	afterEach(() => {
-		if (prevAgentDir === undefined) delete process.env.BOBBIT_AGENT_DIR;
-		else process.env.BOBBIT_AGENT_DIR = prevAgentDir;
-		resetAgentDirStateForTests();
-		fs.rmSync(tmpAgentDir, { recursive: true, force: true });
-	});
-
-	it("emits per-model api/baseUrl for well-known models and never sends reasoning_effort+tools on a forbidden completions model", () => {
-		const models = translateWellKnown(loadFixture(), GATEWAY);
-		writeAigwModelsJson(`${GATEWAY}/v1`, models);
-		const data = JSON.parse(fs.readFileSync(path.join(tmpAgentDir, "models.json"), "utf-8"));
-		const entries: any[] = data.providers.aigw.models;
-
-		const sol = entries.find((m) => m.id === "gpt-5.6-sol");
-		assert.ok(sol, "expected gpt-5.6-sol entry with bare id");
-		assert.equal(sol.api, "openai-responses");
-		assert.equal(sol.upstreamProvider, "openai");
-		assert.equal(sol.baseUrl, "http://aigw-local.t3.zone/openai/v1");
-
-		const claude = entries.find((m) => m.id === "us.anthropic.claude-opus-4-6");
-		assert.equal(claude.api, "bedrock-converse-stream");
-		assert.equal(claude.baseUrl, "http://aigw-local.t3.zone/aws");
-
-		const qwen = entries.find((m) => m.id === "qwen3-coder-480b-a35b");
-		assert.equal(qwen.api, "openai-completions");
-		assert.equal(qwen.baseUrl, "http://aigw-local.t3.zone/gresearch/v1");
-
-		// The forbidden combo (reasoning_effort + tools on plain chat/completions)
-		// cannot occur: every gpt-5.6 model routes to openai-responses, never
-		// openai-completions.
-		const forbidden = entries.filter((m) => m.id.startsWith("gpt-5.6") && m.api === "openai-completions");
-		assert.deepEqual(forbidden, [], "gpt-5.6 models must never be routed to openai-completions");
-	});
-
-	it("normalizes legacy AIGW provider-prefixed ids when the bare well-known id exists", () => {
-		const models = translateWellKnown(loadFixture(), GATEWAY);
-		writeAigwModelsJson(`${GATEWAY}/v1`, models);
-
-		assert.equal(normalizeAigwModelString("aigw/openai/gpt-5.6-sol"), "aigw/gpt-5.6-sol");
-		assert.equal(normalizeAigwModelString("aigw/aws-mantle/gpt-oss-120b"), "aigw/gpt-oss-120b");
-		assert.equal(normalizeAigwModelString("aigw/unknown/gpt-5.6-sol"), "aigw/gpt-5.6-sol");
-		assert.equal(normalizeAigwModelString("aigw/gresearch/not-present"), "aigw/gresearch/not-present");
-		assert.equal(normalizeAigwModelString("aigw/unknown/multi/segment"), "aigw/unknown/multi/segment");
-	});
-
-	it("preserves legacy prefixes while an old models file has duplicate bare ids", () => {
-		fs.writeFileSync(path.join(tmpAgentDir, "models.json"), JSON.stringify({ providers: { aigw: { models: [
-			{ id: "shared", upstreamProvider: "first" },
-			{ id: "shared", upstreamProvider: "second" },
-		] } } }));
-		assert.equal(normalizeAigwModelString("aigw/first/shared"), "aigw/first/shared");
-	});
-
-	it("does not persist configuration when the atomic models.json write fails", async () => {
-		const badAgentDir = path.join(tmpAgentDir, "not-a-directory");
-		fs.writeFileSync(badAgentDir, "file");
-		process.env.BOBBIT_AGENT_DIR = badAgentDir;
-		resetAgentDirStateForTests();
-		const prefs = new Map<string, unknown>();
-		const server = await startDiscoveryServer((_req, res) => {
-			res.setHeader("Content-Type", "application/json");
-			res.end(JSON.stringify({ provider: {} }));
-		});
-		try {
-			await assert.rejects(() => configureAigw(server.origin, {
-				get: (key: string) => prefs.get(key),
-				set: (key: string, value: unknown) => { prefs.set(key, value); },
-				remove: (key: string) => { prefs.delete(key); },
-			} as any));
-			assert.equal(prefs.has("aigw.url"), false);
-		} finally {
-			await server.close();
-			process.env.BOBBIT_AGENT_DIR = tmpAgentDir;
-			resetAgentDirStateForTests();
-		}
-	});
-});
-
-describe("discoverAigwModels — fallback option-1 (no well-known probe under guards)", () => {
-	function startMock(models: Array<{ id: string }>): Promise<{ url: string; close: () => Promise<void>; paths: () => string[] }> {
-		const paths: string[] = [];
-		const server = http.createServer((req, res) => {
-			paths.push(req.url ?? "");
-			if (req.url?.endsWith("/v1/models")) {
-				res.writeHead(200, { "Content-Type": "application/json" });
-				res.end(JSON.stringify({ data: models.map((m) => ({ ...m, object: "model" })) }));
-				return;
-			}
-			res.writeHead(404);
-			res.end();
-		});
-		return new Promise((resolve) => {
-			server.listen(0, "127.0.0.1", () => {
-				const port = (server.address() as any).port;
-				resolve({
-					url: `http://127.0.0.1:${port}`,
-					close: () => new Promise<void>((r) => server.close(() => r())),
-					paths: () => paths,
-				});
-			});
-		});
-	}
-
-	afterEach(() => configureAigwRuntimeFlags({ skipAigwDiscovery: false, testNoExternal: false }));
-
-	it("does not probe /.well-known when skipAigwDiscovery is set, and routes OpenAI reasoning ids to openai-responses while Claude stays non-responses", async () => {
-		configureAigwRuntimeFlags({ skipAigwDiscovery: true });
-		const mock = await startMock([
-			{ id: "openai/gpt-5.2" },
-			{ id: "aws/us.anthropic.claude-sonnet-4-6" },
-			{ id: "gresearch/qwen3-coder" },
-		]);
-		try {
-			const models = await discoverAigwModels(mock.url);
-			assert.ok(!mock.paths().some((p) => p.includes(".well-known")),
-				`no well-known probe expected under skip flag, saw: ${mock.paths().join(",")}`);
-
-			const gpt = models.find((m: any) => (m.wireId ?? m.id) === "gpt-5.2");
-			assert.ok(gpt, "gpt-5.2 should be present with a bare wire id");
-			assert.equal(gpt.api, "openai-responses", "OpenAI reasoning id routes to responses in fallback");
-			assert.ok(gpt.baseUrl?.endsWith("/openai/v1"), `baseUrl should be origin/openai/v1, got ${gpt.baseUrl}`);
-			assert.equal(gpt.wireId, "gpt-5.2");
-			assert.equal(gpt.compat?.supportsReasoningEffort, true, "fallback Responses models must retain reasoning effort");
-
-			// Claude is NOT routed to responses in fallback (remapped to bedrock downstream).
-			const claude = models.find((m: any) => m.id.includes("claude"));
-			assert.ok(claude);
-			assert.notEqual(claude.api, "openai-responses");
-
-			// Non-reasoning model keeps openai-completions.
-			const qwen = models.find((m: any) => m.id === "gresearch/qwen3-coder");
-			assert.ok(qwen, "expected non-reasoning fallback model");
-			assert.equal(qwen.api, "openai-completions");
-			assert.equal(qwen.baseUrl, `${mock.url}/v1`, "legacy completions retain the /v1 route from /v1/models");
-		} finally {
-			await mock.close();
-		}
-	});
-
-	it("blocks external well-known + models probes under testNoExternal", async () => {
-		configureAigwRuntimeFlags({ testNoExternal: true });
-		await assert.rejects(
-			() => discoverAigwModels("https://api.example.invalid/v1"),
-			/External AI Gateway discovery is disabled in tests/,
-		);
 	});
 });
 
@@ -395,239 +193,5 @@ describe("well-known collision, validation, and pricing invariants", () => {
 		const wellKnown = normalizeWellKnownCost({ input: 2, output: 3 });
 		assert.equal(legacy.input, wellKnown.input * 1_000_000);
 		assert.equal(legacy.output, wellKnown.output * 1_000_000);
-	});
-});
-
-function startDiscoveryServer(
-	handler: (req: http.IncomingMessage, res: http.ServerResponse, origin: string) => void,
-): Promise<{ origin: string; close: () => Promise<void> }> {
-	const server = http.createServer((req, res) => handler(req, res, `http://127.0.0.1:${(server.address() as any).port}`));
-	return new Promise((resolve) => server.listen(0, "127.0.0.1", () => resolve({
-		origin: `http://127.0.0.1:${(server.address() as any).port}`,
-		close: () => new Promise<void>((done) => server.close(() => done())),
-	})));
-}
-
-describe("AIGW DNS guard lifecycle", () => {
-	afterEach(() => replaceAigwProviderDnsGuardHosts([]));
-
-	it("does not activate rejected or merely translated provider hosts", async () => {
-		replaceAigwProviderDnsGuardHosts(["active.example"]);
-		translateWellKnown({ provider: { candidate: {
-			npm: "@ai-sdk/openai", options: { baseURL: "https://candidate.example/v1" }, models: { model: {} },
-		} } }, GATEWAY);
-		assert.deepEqual(getAigwProviderDnsGuardHosts(), ["active.example"]);
-
-		const privateLookup = ((_hostname: string, _options: any, callback: any) => callback(null, [
-			{ address: "169.254.169.254", family: 4 },
-		])) as any;
-		const admitted = await filterValidatedProviderUrls({ provider: { candidate: {
-			options: { baseURL: "https://candidate.example/v1" }, models: { model: {} },
-		} } }, new URL(GATEWAY).origin, Date.now() + 250, privateLookup);
-		assert.deepEqual(admitted.provider, {});
-		assert.deepEqual(getAigwProviderDnsGuardHosts(), ["active.example"], "rejected discovery must not change active DNS behavior");
-	});
-
-	it("replaces and clears the active host set on removal", () => {
-		const agentDir = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-aigw-remove-"));
-		const previousAgentDir = process.env.BOBBIT_AGENT_DIR;
-		try {
-			process.env.BOBBIT_AGENT_DIR = agentDir;
-			resetAgentDirStateForTests();
-			replaceAigwProviderDnsGuardHosts(["old.example"]);
-			replaceAigwProviderDnsGuardHosts(["new.example"]);
-			assert.deepEqual(getAigwProviderDnsGuardHosts(), ["new.example"]);
-			removeAigw({ remove() {} } as any);
-			assert.deepEqual(getAigwProviderDnsGuardHosts(), []);
-		} finally {
-			if (previousAgentDir === undefined) delete process.env.BOBBIT_AGENT_DIR;
-			else process.env.BOBBIT_AGENT_DIR = previousAgentDir;
-			resetAgentDirStateForTests();
-			fs.rmSync(agentDir, { recursive: true, force: true });
-		}
-	});
-
-	it("bounds and consolidates provider DNS admission by the shared deadline", async () => {
-		let lookups = 0;
-		const hangingLookup = ((_hostname: string, _options: any, _callback: any) => { lookups++; }) as any;
-		const config = {
-			disabled_providers: ["disabled"],
-			provider: {
-				first: { options: { baseURL: "https://slow.example/one" }, models: {} },
-				second: { options: { baseURL: "https://slow.example/two" }, models: {} },
-				disabled: { options: { baseURL: "https://disabled-slow.example/v1" }, models: {} },
-			},
-		};
-		const started = Date.now();
-		const admitted = await filterValidatedProviderUrls(config, new URL(GATEWAY).origin, Date.now() + 40, hangingLookup);
-		assert.deepEqual(admitted.provider, {});
-		assert.equal(lookups, 1, "duplicate provider hostnames share one DNS admission");
-		assert.ok(Date.now() - started < 500, "DNS admission must stop at the shared deadline");
-	});
-
-	it("executes the generated guard for public/private lookups and real requests", async () => {
-		const root = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-aigw-guard-"));
-		const previousAgentDir = process.env.BOBBIT_AGENT_DIR;
-		const previousBobbitDir = process.env.BOBBIT_DIR;
-		const originalLookup = dnsCallback.lookup;
-		let answers: Array<{ address: string; family: number }> = [];
-		try {
-			process.env.BOBBIT_AGENT_DIR = path.join(root, "agent");
-			process.env.BOBBIT_DIR = path.join(root, "headquarters");
-			resetAgentDirStateForTests();
-			writeAigwModelsJson(GATEWAY, [{
-				id: "model", wireId: "model", name: "Model", api: "openai-responses",
-				baseUrl: "https://api.vendor.example/v1", upstreamProvider: "vendor",
-				reasoning: false, input: ["text"], contextWindow: 1, maxTokens: 1,
-				cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-			}]);
-			const extension = writeAigwDnsGuardExtension();
-			assert.ok(extension && fs.existsSync(extension));
-			assert.match(extension!, /aigw-dns-guard/);
-
-			// The generated module captures this resolver as its connection-time
-			// source, then installs its own dns.lookup wrapper on import.
-			dnsCallback.lookup = ((_hostname: string, options: any, callback?: any) => {
-				const cb = typeof options === "function" ? options : callback;
-				cb(null, answers);
-			}) as any;
-			const loaded = await import(`${pathToFileURL(extension!).href}?test=${Date.now()}`);
-			assert.equal(typeof loaded.default, "function");
-
-			const guardedLookup = (options: any = { all: true }) => new Promise<any[]>((resolve, reject) => {
-				dnsCallback.lookup("api.vendor.example", options, (error: Error | null, result: any) => error ? reject(error) : resolve(result));
-			});
-			answers = [{ address: "93.184.216.34", family: 4 }];
-			assert.deepEqual(await guardedLookup(), answers, "generated extension must pass public answers to the socket caller");
-
-			answers = [
-				{ address: "93.184.216.34", family: 4 },
-				{ address: "169.254.169.254", family: 4 },
-			];
-			await assert.rejects(guardedLookup(), /non-public address/,
-				"a mixed answer set must fail closed rather than selecting its public member");
-
-			// Exercise Node's real HTTP client path, not just a direct helper call:
-			// request socket creation must stop at the generated lookup guard.
-			answers = [{ address: "127.0.0.1", family: 4 }];
-			await assert.rejects(new Promise<void>((resolve, reject) => {
-				const request = http.get("http://api.vendor.example/provider-probe", () => resolve());
-				request.once("error", reject);
-				request.setTimeout(1_000, () => request.destroy(new Error("request unexpectedly reached socket timeout")));
-			}), /non-public address/);
-		} finally {
-			dnsCallback.lookup = originalLookup;
-			if (previousAgentDir === undefined) delete process.env.BOBBIT_AGENT_DIR;
-			else process.env.BOBBIT_AGENT_DIR = previousAgentDir;
-			if (previousBobbitDir === undefined) delete process.env.BOBBIT_DIR;
-			else process.env.BOBBIT_DIR = previousBobbitDir;
-			resetAgentDirStateForTests();
-			fs.rmSync(root, { recursive: true, force: true });
-		}
-	});
-});
-
-describe("well-known bounded one-hop resolution", () => {
-	afterEach(() => {
-		configureAigwRuntimeFlags({ skipAigwDiscovery: false, testNoExternal: false, e2e: false });
-		delete process.env.AIGW_OPENCODE_TOKEN;
-	});
-
-	it("resolves one same-origin remote hop within two requests and filters declared headers", async () => {
-		process.env.AIGW_OPENCODE_TOKEN = "inherited-secret";
-		const paths: string[] = [];
-		const auth: Array<string | undefined> = [];
-		const hosts: Array<string | undefined> = [];
-		const server = await startDiscoveryServer((req, res, origin) => {
-			paths.push(req.url ?? "");
-			auth.push(req.headers.authorization);
-			hosts.push(req.headers.host);
-			res.setHeader("Content-Type", "application/json");
-			if (req.url === "/.well-known/opencode") {
-				res.end(JSON.stringify({ remote_config: { url: `${origin}/remote`, headers: {
-					Authorization: "Bearer declared", Host: "evil.invalid", "Content-Length": "999", "User-Agent": "evil",
-				} } }));
-			} else {
-				res.end(JSON.stringify({ config: { provider: {} } }));
-			}
-		});
-		try {
-			const config = await fetchWellKnownConfig(server.origin);
-			assert.deepEqual(config?.provider, {});
-			assert.deepEqual(paths, ["/.well-known/opencode", "/remote"]);
-			assert.equal(auth[0], "Bearer inherited-secret");
-			assert.equal(auth[1], "Bearer declared");
-			assert.ok(hosts.every((host) => host === new URL(server.origin).host));
-		} finally { await server.close(); }
-	});
-
-	it("rejects a second remote hop without requesting it", async () => {
-		const paths: string[] = [];
-		const server = await startDiscoveryServer((req, res, origin) => {
-			paths.push(req.url ?? "");
-			res.setHeader("Content-Type", "application/json");
-			res.end(JSON.stringify(req.url === "/.well-known/opencode"
-				? { remote_config: { url: `${origin}/one` } }
-				: { remote_config: { url: `${origin}/two` } }));
-		});
-		try {
-			assert.equal(await fetchWellKnownConfig(server.origin), null);
-			assert.deepEqual(paths, ["/.well-known/opencode", "/one"]);
-		} finally { await server.close(); }
-	});
-
-	it("does not follow redirects and rejects cross-origin HTTP remote targets", async () => {
-		const redirectPaths: string[] = [];
-		const redirecting = await startDiscoveryServer((req, res, origin) => {
-			redirectPaths.push(req.url ?? "");
-			if (req.url === "/.well-known/opencode") {
-				res.setHeader("Content-Type", "application/json");
-				res.end(JSON.stringify({ remote_config: { url: `${origin}/redirect` } }));
-			} else {
-				res.writeHead(302, { Location: "/config" });
-				res.end();
-			}
-		});
-		try {
-			assert.equal(await fetchWellKnownConfig(redirecting.origin), null);
-			assert.deepEqual(redirectPaths, ["/.well-known/opencode", "/redirect"]);
-		} finally { await redirecting.close(); }
-
-		let remoteRequests = 0;
-		const remote = await startDiscoveryServer((_req, res) => { remoteRequests++; res.end("{}"); });
-		const gateway = await startDiscoveryServer((_req, res) => {
-			res.setHeader("Content-Type", "application/json");
-			res.end(JSON.stringify({ remote_config: { url: `${remote.origin}/config` } }));
-		});
-		try {
-			assert.equal(await fetchWellKnownConfig(gateway.origin), null);
-			assert.equal(remoteRequests, 0);
-		} finally { await gateway.close(); await remote.close(); }
-	});
-
-	it("enforces an absolute deadline even when the server keeps streaming bytes", async () => {
-		const server = await startDiscoveryServer((_req, res) => {
-			res.writeHead(200, { "Content-Type": "application/json" });
-			const timer = setInterval(() => res.write(" "), 15);
-			res.on("close", () => clearInterval(timer));
-		});
-		try {
-			const started = Date.now();
-			assert.equal(await fetchWellKnownConfig(server.origin, 100), null);
-			assert.ok(Date.now() - started < 1000, "streaming response must not extend the end-to-end deadline");
-		} finally { await server.close(); }
-	});
-
-	it("treats an empty provider object as authoritative and never calls /v1/models", async () => {
-		const paths: string[] = [];
-		const server = await startDiscoveryServer((req, res) => {
-			paths.push(req.url ?? "");
-			res.setHeader("Content-Type", "application/json");
-			res.end(JSON.stringify({ provider: {} }));
-		});
-		try {
-			assert.deepEqual(await discoverAigwModels(server.origin), []);
-			assert.deepEqual(paths, ["/.well-known/opencode"]);
-		} finally { await server.close(); }
 	});
 });
